Added bookmark support for multiple Rokus/frontends
Users now have the ability to pause on one Roku/frontend and resume on
another Roku/frontend. Note that the Roku will automatically save the
bookmarks every 30 seconds, however, users will need to manually save
the bookmarks on a frontend.

There are some additional notes about calcuating the bookmarks in the
README file.